requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-environ==0.11.2
django-rest-framework==0.1.0
django-unfold==0.43.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.2
psycopg==3.2.3
psycopg-binary==3.2.3
PyJWT==2.10.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
